Camera Turn sound implemented, turn behavior sharper
Camera Turn:
The gameObject "cameraTurn" now has an AudioSource and the script CameraTurn has two Clips (one for clockwise and one for counter clockwise).
A rather convoluted switch method determines the turn direction of the camera upon pressure plate activation.
Depending on the turn direction one of the two sounds is played.
Stepping on the pressure plate of the front-facing platform no longer registers as an activation and subsequent 0-difference-turn.
When the turn timer runs out, the rotation of the camera is SET to its target angle (this results in less floaty starting angles for the next rotation)
Additionally, those last two changes now prevent those weird turns of <5 degrees when stepping on the front facing pressure plate.