Coin Spawn behavior changes
Picking up coins now starts their respective platform's descend procedures (most of the time)
This tends to lead to a three-platform "grasing" and a subsequent delay before the next three platforms spawn.
This delay forces a certain amount of player combat.
Speaking of delay: There is a 2 second timer before the next coin spawns... this, however, doesn't seem to register.

Coin Mesh is now scaled to 1.5 across the board.

Various cosmetic changes to the scripts and a bit of out-commenting of superfluous code.